small changes ... about to refactor

+++ workHorse.py
@@ -12,23 +12,23 @@ class workHorse:
     def __init__(self, filename, trace):
         self.filename = filename
         self.trace = trace
-        os.remove(self.trace)
-        self.m2dir = expanduser('~')
-        self.m2dir += '/.m2/repository/'
+        self.m2dir = expanduser('~') + '/.m2/repository/'
         self.parent_data = []
         self.parent_data2 = []
+        if os.path.isfile(self.trace):
+            os.remove(self.trace)
 
     def currentPomAttributes(self):
         try:
             fh = open(self.filename, 'r')
         except OSError:
-            print("Oops that was not supposed to happen. Bye...")
+            print("Oops that was not supposed to happen. Bye..." + str(OSError.errno))
             sys.exit(1)
 
         try:
             self.wh = open(self.trace, 'x')
         except OSError:
-            print("Oops that was not supposed to happen. Bye...")
+            print("Oops that was not supposed to happen. Bye..." + str(OSError.errno))
             sys.exit(1)
 
         soup = BeautifulSoup(fh, "xml")
@@ -39,6 +39,7 @@ class workHorse:
         name_list = soup.find_all('name')
         version_list = soup.find_all('version')
         relativePath_list = soup.find_all('relativePath')
+        module_list = soup.find_all('modules')
 
         self.checkListForParent(groupId_list)
         self.checkListForParent(artifactId_list)
@@ -46,23 +47,22 @@ class workHorse:
         self.checkListForParent(name_list)
         self.checkListForParent(version_list)
         self.checkListForParent(relativePath_list)
+        self.checkListForParent(module_list)
 
         self.extract_parent_data(self.parent_data)
+
         for i in self.parent_data2:
-            #print(str(i.items()))
             print(str(i.keys()) + '\n')
             print(str(i.values()) + '\n')
-        #parent_data2.items
-        self.wh.close()
 
-    def __currentPomAttributes(self):
-        print('internal wrapped method')
+        #self.parent_data2.
+
+        self.wh.close()
 
     def checkListForParent(self, the_list):
         for element in the_list:
             parent_info = element.find_parent()
             if parent_info.name == 'parent':
-                #self.save_for_next_iteration(element)
                 self.parent_data.append(element)
             else:
                 self.writeToFile(element)
@@ -80,10 +80,6 @@ class workHorse:
 
     def extract_parent_data(self, listOfTagsFromParent):
         parentDict = {}
-        #print(str(type(listOfTagsFromParent)))
         for element in listOfTagsFromParent:
-            #print(str(type(element.name)))
-            #print(str(type(str(element.string))))
-            #print(str(type(element.tag)))
             parentDict.update({element.name: element.string})
         self.parent_data2.append(parentDict)
